backend/pyproject.toml
[tool.poetry]
name = "asha-transcription"
version = "0.1.0"
description = "Asha async transcription service using FastAPI and WebSockets"
authors = ["Asha"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.40"
pydantic = "^2.11.2"
loguru = "^0.7.3"
httpx = "^0.28.1"
python-multipart = "^0.0.20"
websockets = "^15.0.1"
aiosqlite = "^0.21.0"
greenlet = "^3.1.1"
pydantic-settings = "^2.8.1"
pydub = "^0.25.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
